core/app/requirements.txt
fastapi
uvicorn
requests
python-dotenv
httpx
qdrant-client
pyyaml
cryptography
eth-account
python-gnupg
embit
